tool/requirements.txt
Pillow==10.3.0
playwright==1.43.0
Requests==2.31.0
